zotero_topic_modeling/requirements.txt
# Core dependencies
pyzotero==1.5.1
pdfminer.six==20221105
gensim==4.3.2
nltk==3.8.1
matplotlib==3.8.2
numpy==1.26.2
requests==2.31.0

# UI dependencies
tk==0.1.0  # Usually comes with Python, version may vary

# Optional but recommended for development
# pytest==7.4.3
# pylint==3.0.2
# black==23.11.0

# Note: Specific versions are listed for reproducibility.
# These versions have been tested and work together.
# Newer versions might work but haven't been verified.
